#119 [feat] : 불필요한 route 제거, 공지사항 상세 인 경우 버튼 분기처리
